backend/requirements.txt
getdaft[aws]
awscli-local[ver1]
Pillow
opencv-python
soundfile
numpy
